Movie API
I will be creating a Server Side API that will include different elements.
The API must use Node.js , REST architecture, with URL endpoints.
There must be at least three middleware modules. 
It also must have a package.json.
The database will be built with MongoDB and use business logic made with Mongoose.
The API will provide movie information in JSON format and the JavaScript must be error free.
The API will tested in Postman. 
There need to be Authentication and Authorization code as well as  data validation logic so it can meet security regulations. 
The API will be deployed to Heroku so it will be public.